Split project into multiple files to help avoid merge errors
